+++ Runtime/Scripts/Extensions/Integrals/Int/IntExtensions.Pow.cs
@@ -6,7 +6,7 @@ namespace NumericMath
 
 	public static partial class IntExtensions
 	{
-		public static int Pow(this int value, int exponent)
+		public static int Pow(this int value, double exponent)
 		{
 			return (int)Math.Pow(value, exponent);
 		}
